Extract HTTP href discovery
