+++ src/Http/HttpResource.php
@@ -12,7 +12,6 @@ use BEAR\Resource\Uri as ResourceUri;
 use Koriym\PhpServer\PhpServer;
 use LogicException;
 use Override;
-use stdClass;
 
 use function curl_exec;
 use function curl_init;
@@ -21,18 +20,12 @@ use function escapeshellarg;
 use function explode;
 use function file_exists;
 use function file_put_contents;
-use function html_entity_decode;
 use function http_build_query;
 use function implode;
 use function in_array;
-use function is_array;
 use function is_string;
 use function json_encode;
-use function preg_match;
-use function preg_match_all;
-use function preg_split;
 use function sprintf;
-use function strtolower;
 use function strtoupper;
 use function trim;
 
@@ -41,8 +34,6 @@ use const CURLOPT_HEADER;
 use const CURLOPT_HTTPHEADER;
 use const CURLOPT_POSTFIELDS;
 use const CURLOPT_RETURNTRANSFER;
-use const ENT_HTML5;
-use const ENT_QUOTES;
 use const FILE_APPEND;
 use const JSON_THROW_ON_ERROR;
 use const PHP_EOL;
@@ -54,6 +45,7 @@ final class HttpResource implements ResourceInterface
     private static PhpServer $server;
     private readonly QueryMerger $queryMerger;
     private readonly CreateResponse $createResponse;
+    private readonly HrefExtractor $hrefExtractor;
 
     public function __construct(string $host, string $index, string $logFile = 'php://stderr')
     {
@@ -64,6 +56,7 @@ final class HttpResource implements ResourceInterface
         $this->startServer($host, $index);
         $this->queryMerger = new QueryMerger();
         $this->createResponse = new CreateResponse();
+        $this->hrefExtractor = new HrefExtractor();
     }
 
     private function startServer(string $host, string $index): void
@@ -137,9 +130,7 @@ final class HttpResource implements ResourceInterface
             throw new LogicException('ResourceObject is required to follow a link.');
         }
 
-        $href = $this->halHref($rel, $ro)
-            ?? $this->semanticHtmlHref($rel, $ro->view)
-            ?? $this->linkHeaderHref($rel, $ro->headers);
+        $href = $this->hrefExtractor->href($rel, $ro);
         if ($href === null) {
             throw new LogicException(sprintf('Link rel `%s` is not available.', $rel));
         }
@@ -239,143 +230,6 @@ final class HttpResource implements ResourceInterface
         $curlCommand = sprintf('curl -s -i %s', escapeshellarg($url));
 
         return $this->request($url, 'get', $curlCommand);
-    }
-
-    private function halHref(string $rel, ResourceObject $ro): string|null
-    {
-        $body = is_array($ro->body) ? $ro->body : [];
-        $links = $body['_links'] ?? null;
-        if ($links instanceof stdClass) {
-            $links = (array) $links;
-        }
-
-        if (! is_array($links) || ! isset($links[$rel])) {
-            return null;
-        }
-
-        $link = $links[$rel];
-        if ($link instanceof stdClass) {
-            $link = (array) $link;
-        }
-
-        if (! is_array($link)) {
-            return null;
-        }
-
-        $href = $link['href'] ?? null;
-
-        return is_string($href) ? $href : null;
-    }
-
-    private function semanticHtmlHref(string $rel, string|null $view): string|null
-    {
-        if (! is_string($view) || $view === '') {
-            return null;
-        }
-
-        if (preg_match_all('/<(?:a|area|link)\b(?P<attrs>[^>]*)>/i', $view, $matches) !== 1) {
-            return null;
-        }
-
-        foreach ($matches['attrs'] as $attrs) {
-            if (! $this->hasLinkToken($attrs, $rel)) {
-                continue;
-            }
-
-            $href = $this->attribute($attrs, 'href');
-            if ($href !== null && $href !== '') {
-                return $href;
-            }
-        }
-
-        return null;
-    }
-
-    /** @param array<string, mixed> $headers */
-    private function linkHeaderHref(string $rel, array $headers): string|null
-    {
-        $header = $this->headerValue($headers, 'Link');
-        if ($header === null) {
-            return null;
-        }
-
-        $links = preg_split('/,\s*(?=<)/', $header);
-        if (! is_array($links)) {
-            return null;
-        }
-
-        foreach ($links as $link) {
-            if (preg_match('/^\s*<([^>]*)>\s*(.*)$/', $link, $match) !== 1) {
-                continue;
-            }
-
-            $linkRel = $this->linkHeaderParam($match[2], 'rel');
-            if ($linkRel === null || ! $this->containsToken($linkRel, $rel)) {
-                continue;
-            }
-
-            return html_entity_decode($match[1], ENT_QUOTES | ENT_HTML5);
-        }
-
-        return null;
-    }
-
-    private function hasLinkToken(string $attrs, string $rel): bool
-    {
-        $relAttr = $this->attribute($attrs, 'rel');
-        if ($relAttr !== null && $this->containsToken($relAttr, $rel)) {
-            return true;
-        }
-
-        $classAttr = $this->attribute($attrs, 'class');
-
-        return $classAttr !== null && $this->containsToken($classAttr, $rel);
-    }
-
-    private function attribute(string $attrs, string $name): string|null
-    {
-        if (preg_match('/\b' . $name . '\s*=\s*(?:"([^"]*)"|\'([^\']*)\'|([^\s>]+))/i', $attrs, $match) !== 1) {
-            return null;
-        }
-
-        $value = $match[1] ?? $match[2] ?? $match[3] ?? '';
-
-        return html_entity_decode($value, ENT_QUOTES | ENT_HTML5);
-    }
-
-    private function linkHeaderParam(string $attrs, string $name): string|null
-    {
-        if (preg_match('/(?:^|;)\s*' . $name . '\s*=\s*(?:"([^"]*)"|([^;\s]+))/i', $attrs, $match) !== 1) {
-            return null;
-        }
-
-        return $match[1] ?? $match[2] ?? null;
-    }
-
-    private function containsToken(string $value, string $token): bool
-    {
-        $tokens = preg_split('/\s+/', trim($value));
-        if (! is_array($tokens)) {
-            return false;
-        }
-
-        return in_array($token, $tokens, true);
-    }
-
-    /** @param array<string, mixed> $headers */
-    private function headerValue(array $headers, string $name): string|null
-    {
-        foreach ($headers as $header => $value) {
-            if (! is_string($value)) {
-                continue;
-            }
-
-            if (strtolower($header) === strtolower($name)) {
-                return $value;
-            }
-        }
-
-        return null;
     }
 
     /** @param array<string, mixed> $query */
